refactor(connector): publish local MCP server
